+++ client/Packages/com.beamable/Editor/Modules/Content/ContentIO.cs
@@ -947,7 +947,7 @@ namespace Beamable.Editor.Content
 
 			var api = await EditorAPI.Instance;
 			var allContent = api.ContentIO.FindAll();
-
+			
 			List<ContentObject> contentList = null;
 			if (allContent != null)
 			{
@@ -960,32 +960,37 @@ namespace Beamable.Editor.Content
 				return;
 			}
 
-			// check for local changes
+			// get all valid (up-to-date) content pieces
+			List<ContentObject> objectsToBake = new List<ContentObject>();
 			foreach (var content in contentList)
 			{
 				var status = await api.ContentIO.GetStatus(content);
-				if (status != ContentStatus.CURRENT)
+				if (status == ContentStatus.CURRENT)
 				{
-					bool result = EditorUtility.DisplayDialog("Local changes",
-						"You have local changes in your content. Do you want to proceed with baking using this data?",
-						"Yes", "No");
-
-					if (!result)
-					{
-						return;
-					}
-
-					break;
+					objectsToBake.Add(content);
 				}
 			}
 
-			BakeLog($"Baking {contentList.Count} items");
+			// check for local changes
+			if (objectsToBake.Count != contentList.Count)
+			{
+				bool continueBaking = EditorUtility.DisplayDialog("Local changes",
+				                                             "You have local changes in your content. " +
+				                                             "Do you want to proceed with baking using only the unchanged data?",
+				                                             "Yes", "No");
+				if (!continueBaking)
+				{
+					return;
+				}
+			}
+			
+			BakeLog($"Baking {objectsToBake.Count} items");
 
 			var serverManifest = await api.ContentIO.FetchManifest();
 
 			bool compress = ContentConfiguration.Instance.EnableBakedContentCompression;
 
-			if (Bake(contentList, serverManifest, compress, out int objectsBaked))
+			if (Bake(objectsToBake, serverManifest, compress, out int objectsBaked))
 			{
 				BakeLog($"Baked {objectsBaked} content objects to '{ContentConstants.BakedContentFilePath + ".bytes"}'");
 			}
@@ -1007,9 +1012,8 @@ namespace Beamable.Editor.Content
 				var serverReference = serverManifest.References.Find(reference => reference.Id == content.Id);
 				if (serverReference == null)
 				{
-					// this content object exists only locally
-					objectsBaked--;
-					continue;
+					throw new Exception($"Content object with ID {content.Id} is missing in a remote manifest." +
+										"Reset your content and try again.");
 				}
 				var version = serverReference.Version;
 				content.SetIdAndVersion(content.Id, version);