+++ client/Packages/com.beamable/Editor/UI/Buss/ThemeManager/BussPropertyVisualElements/BussPropertyVisualElement.cs
@@ -1,5 +1,4 @@
-﻿using Beamable.Editor.UI.Buss;
-using Beamable.Editor.UI.Common;
+﻿using Beamable.Editor.UI.Common;
 using Beamable.UI.Buss;
 using System;
 using UnityEditor;