+++ client/Packages/com.beamable/Editor/UI/Common/Components/ManifestButtonVisualElement/ManifestButtonVisualElement.cs
@@ -1,10 +1,6 @@
 using Beamable.Content;
-using Beamable.Editor.Content;
-using Beamable.Editor.UI.Buss;
-using Beamable.Editor.UI.Buss.Components;
 using Beamable.Editor.UI.Common.Models;
 using System.Collections.Generic;
-using System.Linq;
 using UnityEditor;
 using UnityEngine;
 #if UNITY_2018