+++ client/Packages/com.beamable/Editor/UI/Common/Components/UpdateAvailableAnnouncementVisualElement/UpdateAvailableAnnouncementVisualElement.cs
@@ -1,8 +1,6 @@
 using Beamable.Editor.Environment;
 using Beamable.Editor.Toolbox.Models;
-using Beamable.Editor.Toolbox.UI.Components;
-using Beamable.Editor.UI.Buss;
-using UnityEngine;
+using Beamable.Editor.UI.Components;
 #if UNITY_2018
 using UnityEngine.Experimental.UIElements;
 using UnityEditor.Experimental.UIElements;